cafa_6_protein_function_prediction/requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine learning
scikit-learn>=1.0.0
torch>=2.0.0
transformers>=4.30.0

# Data processing
biopython>=1.79

# Configuration management
omegaconf>=2.3.0
pyyaml>=6.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.12.0

# Utilities
tqdm>=4.62.0
joblib>=1.1.0

# Kaggle API
kaggle>=1.5.12

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Code quality
black>=22.0.0
flake8>=4.0.0
isort>=5.10.0
mypy>=0.950

# Optional: Lightning for advanced training
pytorch-lightning>=2.0.0

# Optional: SHAP for interpretability
shap>=0.41.0

